# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ModaBrasileira",
        "summary": "Routes of app",
        "description": "API under development",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Jhonattan Rocha da Silva",
            "url": "http://www.example.com/contact/",
            "email": "jhonattab246rocha@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "identifier": "MIT"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/crud/categories/": {
            "post": {
                "summary": "Create Category",
                "operationId": "create_category_crud_categories__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryBase"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Read Categories",
                "operationId": "read_categories_crud_categories__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "filters",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Filters"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    },
                                    "title": "Response Read Categories Crud Categories  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/crud/categories/{category_id}": {
            "get": {
                "summary": "Read Category",
                "operationId": "read_category_crud_categories__category_id__get",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Category Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Category",
                "operationId": "update_category_crud_categories__category_id__put",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Category Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Category",
                "operationId": "delete_category_crud_categories__category_id__delete",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Category Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/crud/companies": {
            "post": {
                "summary": "Create Company",
                "operationId": "create_company_crud_companies_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyBase"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Read Company",
                "operationId": "read_company_crud_companies_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "filters",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Filters"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    },
                                    "title": "Response Read Company Crud Companies Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/crud/companies/{company_id}": {
            "get": {
                "summary": "Read Company",
                "operationId": "read_company_crud_companies__company_id__get",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Company",
                "operationId": "update_company_crud_companies__company_id__put",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Company Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Company",
                "operationId": "delete_company_crud_companies__company_id__delete",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "subcategories": {
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/SubCategory"
                                },
                                {
                                    "type": "null"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Subcategories"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id",
                    "subcategories"
                ],
                "title": "Category"
            },
            "CategoryBase": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "CategoryBase"
            },
            "CategoryCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id"
                ],
                "title": "CategoryCreate"
            },
            "Company": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "social_reason": {
                        "type": "string",
                        "title": "Social Reason"
                    },
                    "cnpj": {
                        "type": "string",
                        "title": "Cnpj"
                    },
                    "tel": {
                        "type": "string",
                        "title": "Tel"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "address": {
                        "type": "string",
                        "title": "Address"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "employees": {
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/Employees"
                                },
                                {
                                    "type": "null"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Employees"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "social_reason",
                    "cnpj",
                    "tel",
                    "email",
                    "address",
                    "id",
                    "employees"
                ],
                "title": "Company"
            },
            "CompanyBase": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "social_reason": {
                        "type": "string",
                        "title": "Social Reason"
                    },
                    "cnpj": {
                        "type": "string",
                        "title": "Cnpj"
                    },
                    "tel": {
                        "type": "string",
                        "title": "Tel"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "address": {
                        "type": "string",
                        "title": "Address"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "social_reason",
                    "cnpj",
                    "tel",
                    "email",
                    "address"
                ],
                "title": "CompanyBase"
            },
            "CompanyCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "social_reason": {
                        "type": "string",
                        "title": "Social Reason"
                    },
                    "cnpj": {
                        "type": "string",
                        "title": "Cnpj"
                    },
                    "tel": {
                        "type": "string",
                        "title": "Tel"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "address": {
                        "type": "string",
                        "title": "Address"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "social_reason",
                    "cnpj",
                    "tel",
                    "email",
                    "address",
                    "id"
                ],
                "title": "CompanyCreate"
            },
...
            "Employees": {
                "properties": {
                    "id_emp": {
                        "type": "integer",
                        "title": "Id Emp"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "cpf": {
                        "type": "string",
                        "title": "Cpf"
                    },
                    "tel": {
                        "type": "string",
                        "title": "Tel"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "address": {
                        "type": "string",
                        "title": "Address"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "id_emp",
                    "name",
                    "cpf",
                    "tel",
                    "email",
                    "address",
                    "id"
                ],
                "title": "Employees"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SubCategory": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "id_cat": {
                        "type": "integer",
                        "title": "Id Cat"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id_cat",
                    "id"
                ],
                "title": "SubCategory"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/crud/token"
                    }
                }
            }
        }
    }
}